back/requirements.txt
amqp==2.4.2
asgiref==3.6.0
Babel==2.7.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.3.0
certifi==2019.3.9
cffi==1.15.0
chardet==3.0.4
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
defusedxml==0.6.0
distlib==0.3.6
Django==4.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-headers==3.14.0
django-environ==0.10.0
django-geojson==4.0.0
django-rosetta==0.9.0
django-simple-history==2.7.3
django-storages==1.7.1
django-timezone-field==3.0
djangorestframework==3.14.0
docutils==0.14
drf-yasg==1.15.0
et-xmlfile==1.0.1
filelock==3.11.0
flower==0.9.3
geojson==3.0.1
gunicorn==20.1.0
idna==2.8
imagesize==1.1.0
importlib-metadata==6.0.0
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
jsonfield==3.1.0
kombu==4.5.0
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.23.0
openapi-codec==1.3.2
openpyxl==2.6.2
platformdirs==3.2.0
pluggy==0.12.0
polib==1.1.0
psycopg2-binary==2.9.6
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.2
Pyphen==0.9.5
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.3
tinycss2==1.0.2
tornado==5.1.1
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
virtualenv==20.21.0
wcwidth==0.1.7
webencodings==0.5.1
zipp==0.5.2
